+++ lib/change_health/models/claim/submission/other_payer_secondary_identifier.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module ChangeHealth
+  module Models
+    module Claim
+      class OtherPayerSecondaryIdentifier < Model
+        property :qualifier
+        property :identifier
+        property :otherIdentifier, from: :other_identifier
+      end
+    end
+  end
+end