+++ test/change_health/models/claim/submission/claim_information_test.rb
@@ -47,6 +47,62 @@ class ClaimInformationTest < Minitest::Test
                        claim_information.healthCareCodeInformation.first[:diagnosisTypeCode])
         end
       end
+
+      describe 'other payer name' do
+        let(:other_payer_name) do
+          ChangeHealth::Models::Claim::OtherPayerName.new(
+            otherPayerClaimControlNumber: '12345',
+            otherPayerIdentifier: '67890',
+            otherPayerIdentifierTypeCode: 'XX',
+            otherPayerOrganizationName: 'Some Payer'
+          )
+        end
+
+        before do
+          claim_information.otherSubscriberInformation = ChangeHealth::Models::Claim::OtherSubscriberInformation.new
+        end
+
+        it 'defaults to no other payer names' do
+          assert_nil(claim_information.otherSubscriberInformation.otherPayerName)
+        end
+
+        it 'can add an other payer name' do
+          claim_information.otherSubscriberInformation.otherPayerName = other_payer_name
+
+          assert_equal(other_payer_name, claim_information.otherSubscriberInformation.otherPayerName)
+          assert_equal(other_payer_name.otherPayerClaimControlNumber,
+                       claim_information.otherSubscriberInformation.otherPayerName.otherPayerClaimControlNumber)
+        end
+
+        it 'can add other payer secondary identifiers' do
+          secondary_identifier1 = ChangeHealth::Models::Claim::OtherPayerSecondaryIdentifier.new(
+            qualifier: '01',
+            identifier: 'ABC123'
+          )
+          secondary_identifier2 = ChangeHealth::Models::Claim::OtherPayerSecondaryIdentifier.new(
+            qualifier: '02',
+            identifier: 'DEF456'
+          )
+
+          other_payer_name.add_other_payer_secondary_identifier(secondary_identifier1)
+          other_payer_name.add_other_payer_secondary_identifier(secondary_identifier2)
+
+          claim_information.otherSubscriberInformation.otherPayerName = other_payer_name
+
+          assert_equal(
+            2,
+            claim_information.otherSubscriberInformation.otherPayerName.otherPayerSecondaryIdentifier.size
+          )
+          assert_equal(
+            secondary_identifier1.qualifier,
+            claim_information.otherSubscriberInformation.otherPayerName.otherPayerSecondaryIdentifier.first.qualifier
+          )
+          assert_equal(
+            secondary_identifier2.qualifier,
+            claim_information.otherSubscriberInformation.otherPayerName.otherPayerSecondaryIdentifier.last.qualifier
+          )
+        end
+      end
     end
   end
 end