+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+# [6.4.0] - 2025-11-03
+
+### Added
+
+* Added support for Other Payer Claim Control Number & Other Payer Secondary Identifier fields for claim submissions.
+
+### Changed
+
+* SubmissionData now returns the request JSON in addition to the API reponse.
+
 # [6.3.0] - 2025-08-18
 
 ### Added
@@ -800,6 +810,7 @@ Added the ability to hit professional claim submission API. For more details, se
 * Authentication
 * Configuration
 
+[6.4.0]: https://github.com/WeInfuse/change_health/compare/v6.3.0...v6.4.0
 [6.3.0]: https://github.com/WeInfuse/change_health/compare/v6.2.0...v6.3.0
 [6.2.0]: https://github.com/WeInfuse/change_health/compare/v6.1.2...v6.2.0
 [6.1.2]: https://github.com/WeInfuse/change_health/compare/v6.1.1...v6.1.2